# docs/openapi.yaml
openapi: 3.0.3
info:
  title: to-do-list API
  version: "1.0.0"
  description: |
    A small task-management API: JWT authentication, per-user tasks, and
    profile management. See README.md for architecture and setup.
servers:
  - url: http://localhost:8080/api/v1
...
paths:
...
  /users:
    get:
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
      tags: [users]
      summary: List every user (admin only)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: One page of users
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /users/{id}:
    parameters:
      - $ref: '#/components/parameters/UserId'
    get:
      tags: [users]
      summary: Get a user (self, or admin for any user)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [users]
      summary: Update a user (self, or admin for any user)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateUserRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [users]
      summary: Delete a user (self, or admin for any user)
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Deleted }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    UserId:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing, invalid or expired token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated, but not allowed to perform this action
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource does not exist (or belongs to another user)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Page:
      type: object
      description: |
        The envelope every list endpoint returns. limit and offset echo what
        was actually applied after clamping, which is not always what was
        asked for: a limit above the configured maximum comes back trimmed to
        it, so a client can page correctly without knowing the server's
        limits in advance.
      required: [items, total, limit, offset]
      properties:
        total:
          type: integer
          format: int32
          minimum: 0
          description: Rows matching the query, ignoring limit and offset.
        limit:
          type: integer
          format: int32
          minimum: 1
          description: Page size actually used.
        offset:
          type: integer
          format: int32
          minimum: 0
...
    UserPage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/User' }
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: >-
            The stable identifier of the failure. Branch on this; the message
            is written for people and may be reworded at any time.
          enum:
            - validation_error
            - bad_request
            - not_found
            - conflict
            - unauthorized
            - invalid_credentials
            - forbidden
            - method_not_allowed
            - payload_too_large
            - rate_limited
            - internal_error
        message: { type: string }
        request_id: { type: string }
        fields:
          type: array
          description: Which fields were rejected. Sent with validation_error only.
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
...
    User:
      type: object
      description: Never includes the password hash - see README "Fixed issues".
      properties:
        id: { type: integer, format: int64 }
        username: { type: string }
        email: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    UpdateUserRequest:
      type: object
      description: Every field is optional; omitted fields are left unchanged.
      properties:
        # Lengths are counted in Unicode characters, not bytes.
        username: { type: string, minLength: 3, maxLength: 50 }
        email: { type: string, format: email }
        # minLength is in characters; maxLength is bcrypt's 72-BYTE limit,
        # so a password of multi-byte characters may be rejected sooner.
        password: { type: string, minLength: 8, maxLength: 72 }
...